Create topLevelGroup if not-cache hit. Fixes #129
We want to add a group to the topLevelGroups array if it was not caused
by a cache hit (at any level recursively -- because this means we
already would have added it), AND only the top-most group (after
all intermediate groups finish.

This regressed in the most recent fix to fileGroups.

Looks like this regressed with the fix to fileGroups. It seems like
the `topLevelGroups` field is a big source of problems, so I put a TODO
to figure out a way to kill it.